Refactors types
